--- a/snmp/tox.ini
+++ b/snmp/tox.ini
@@ -37,4 +37,4 @@ commands =
 [testenv:bench]
 commands =
     pip install -r requirements.in
-    pytest -v {posargs}
+    pytest -v {posargs} --benchmark-only --benchmark-histogram --benchmark-cprofile=tottime