@mgarabed 2020-08-06
There is a lot of great data here, but it's hard to see the direct comparisons you are trying to make given the way the charts are laid out.  

The first chart provides a good comparison of cache/no-cache, but the scale doesn't allow any comparison of the async/no-async benefits.  The next two charts are better because the scale is adjusted, but the async/no-async aren't directly next to each other making the performance gains harder to see easily.