@AlexandreYang 2020-08-06
@mgarabed 

Yeah, the graph is not that easy to read since we have multiple criteria.

Using `concurrent queries` seems quite effective for OID CACHING + 32/64 Oid batching.

It's a bit easier to read if we group them like this:

```
Name (time in ms)                           OPS         

test_oids_cache_bench[3600-10-False]     2.0586 (0.48)  
test_oids_cache_bench[3600-10-True]      3.4633 (0.81)  

test_oids_cache_bench[3600-32-False]     2.8791 (0.67)  
test_oids_cache_bench[3600-32-True]      4.2405 (0.99)  

test_oids_cache_bench[3600-64-False]     2.9973 (0.70)  
test_oids_cache_bench[3600-64-True]      4.2957 (1.0)   

test_oids_cache_bench[3600-128-False]    3.1390 (0.73)  
test_oids_cache_bench[3600-128-True]     3.5860 (0.83)  

test_oids_cache_bench[3600-256-False]    3.2064 (0.75)  
test_oids_cache_bench[3600-256-True]     3.1752 (0.74)  

```