@AlexandreYang 2020-08-07
Another run of the benchmark:

```

------------------------------------------------------------------------------------------------- benchmark: 20 tests --------------------------------------------------------------------------------------------------
Name (time in ms)                                Min                   Max                  Mean              StdDev                Median                 IQR            Outliers     OPS            Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_oids_cache_bench[3600-32-True]         248.6511 (1.0)        305.9268 (1.01)       271.0525 (1.01)      22.4431 (2.84)       264.8333 (1.01)      30.6992 (3.05)          1;0  3.6893 (0.99)          5           1
test_oids_cache_bench[3600-64-True]         250.2821 (1.01)       301.7639 (1.0)        267.5511 (1.0)       19.8568 (2.51)       263.2345 (1.0)       16.2389 (1.61)          1;1  3.7376 (1.0)           5           1
test_oids_cache_bench[3600-128-True]        290.8060 (1.17)       347.5371 (1.15)       309.7013 (1.16)      23.2885 (2.94)       302.8266 (1.15)      31.3109 (3.11)          1;0  3.2289 (0.86)          5           1
test_oids_cache_bench[3600-10-True]         315.2371 (1.27)       335.3977 (1.11)       327.9126 (1.23)       7.9126 (1.0)        328.3993 (1.25)      10.0650 (1.0)           1;0  3.0496 (0.82)          5           1
test_oids_cache_bench[3600-256-False]       318.4196 (1.28)       363.4123 (1.20)       339.3742 (1.27)      16.5724 (2.09)       341.6714 (1.30)      19.5932 (1.95)          2;0  2.9466 (0.79)          5           1
test_oids_cache_bench[3600-256-True]        329.2596 (1.32)       414.7311 (1.37)       356.4610 (1.33)      33.4586 (4.23)       345.4757 (1.31)      26.1391 (2.60)          1;1  2.8054 (0.75)          5           1
test_oids_cache_bench[3600-128-False]       333.5550 (1.34)       377.0084 (1.25)       346.3554 (1.29)      17.6369 (2.23)       341.3691 (1.30)      16.9710 (1.69)          1;0  2.8872 (0.77)          5           1
test_oids_cache_bench[3600-64-False]        357.9784 (1.44)       414.7732 (1.37)       377.2985 (1.41)      21.8974 (2.77)       369.5102 (1.40)      19.6039 (1.95)          1;0  2.6504 (0.71)          5           1
test_oids_cache_bench[3600-32-False]        378.8990 (1.52)       455.4911 (1.51)       403.1831 (1.51)      30.1059 (3.80)       391.6883 (1.49)      25.0418 (2.49)          1;1  2.4803 (0.66)          5           1
test_oids_cache_bench[3600-10-False]        487.9969 (1.96)       539.2374 (1.79)       503.5730 (1.88)      20.3436 (2.57)       496.0618 (1.88)      15.2671 (1.52)          1;1  1.9858 (0.53)          5           1
test_oids_cache_bench[0-256-False]          755.2153 (3.04)       822.0343 (2.72)       786.9995 (2.94)      33.1660 (4.19)       777.9508 (2.96)      64.8871 (6.45)          2;0  1.2706 (0.34)          5           1
test_oids_cache_bench[0-128-False]          834.5174 (3.36)     1,097.2389 (3.64)       965.5126 (3.61)     102.4527 (12.95)      951.4494 (3.61)     155.6054 (15.46)         2;0  1.0357 (0.28)          5           1
test_oids_cache_bench[0-128-True]           839.5282 (3.38)       951.9554 (3.15)       895.5329 (3.35)      52.3893 (6.62)       922.0679 (3.50)      91.6686 (9.11)          3;0  1.1167 (0.30)          5           1
test_oids_cache_bench[0-32-False]           840.0699 (3.38)       889.3774 (2.95)       855.5510 (3.20)      20.4662 (2.59)       848.5281 (3.22)      26.4839 (2.63)          1;0  1.1688 (0.31)          5           1
test_oids_cache_bench[0-256-True]           852.0268 (3.43)       906.7862 (3.00)       879.0199 (3.29)      23.0628 (2.91)       880.0692 (3.34)      40.2110 (4.00)          2;0  1.1376 (0.30)          5           1
test_oids_cache_bench[0-64-True]            870.9750 (3.50)     1,046.2915 (3.47)       965.1087 (3.61)      74.1652 (9.37)       954.5382 (3.63)     127.0875 (12.63)         2;0  1.0362 (0.28)          5           1
test_oids_cache_bench[0-32-True]            994.5551 (4.00)     1,144.9423 (3.79)     1,080.1978 (4.04)      61.8174 (7.81)     1,072.4435 (4.07)      98.4201 (9.78)          2;0  0.9258 (0.25)          5           1
test_oids_cache_bench[0-64-False]         1,029.3544 (4.14)     1,234.8556 (4.09)     1,123.2857 (4.20)      77.9298 (9.85)     1,129.5106 (4.29)     106.4667 (10.58)         2;0  0.8902 (0.24)          5           1
test_oids_cache_bench[0-10-False]         1,055.4654 (4.24)     1,122.8750 (3.72)     1,084.0513 (4.05)      25.9520 (3.28)     1,077.6621 (4.09)      35.8667 (3.56)          2;0  0.9225 (0.25)          5           1
test_oids_cache_bench[0-10-True]          1,176.2165 (4.73)     1,405.2592 (4.66)     1,267.0164 (4.74)     100.7659 (12.73)    1,211.2637 (4.60)     163.9427 (16.29)         1;0  0.7893 (0.21)          5           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```